fix(security/svelte): address PR #139 review — ssrf module, stale async, navigation
- Extract isSafePushEndpoint() into $lib/server/ssrf.ts (dependency-free) so
  intake routes (api/subscribe, api/notify) don't pull web-push into their bundle
- Harden SSRF guard with full IPv6 coverage: link-local fe80::/10, unique-local
  fc00::/7, IPv4-mapped ::ffff:, IPv6 loopback ::1 (all bracket-notation aware)
- Add cancelled flag to fillNotifState $effect — prevents stale SW registration
  from overwriting state after the user navigates to a different pothole
- Move hitSubmitted + watch state from onMount to $effect so both reset correctly
  on client-side navigation between detail pages; removes onMount import entirely
- Scope getByRole('columnheader') assertions to table in stats.spec.ts — prevents
  false positives from headers in other tables on the page
- Update stale comment in api/ccc: "onMount" → "$effect"
