+++ src/lib/server/webpush.ts
@@ -3,30 +3,7 @@ import { env } from '$env/dynamic/private';
 import { PUBLIC_SUPABASE_URL } from '$env/static/public';
 import { createClient } from '@supabase/supabase-js';
 import { logError } from '$lib/server/observability';
-
-/**
- * Returns true only for HTTPS endpoints that are not loopback, link-local,
- * or private-network addresses. Used before storing or dispatching a push
- * endpoint to prevent SSRF.
- */
-export function isSafePushEndpoint(endpoint: string): boolean {
-	let url: URL;
-	try {
-		url = new URL(endpoint);
-	} catch {
-		return false;
-	}
-	if (url.protocol !== 'https:') return false;
-	const h = url.hostname.toLowerCase();
-	if (h === 'localhost') return false;
-	if (/^127\./.test(h)) return false;
-	if (h === '::1' || h === '[::1]') return false;
-	if (/^10\./.test(h)) return false;
-	if (/^172\.(1[6-9]|2\d|3[01])\./.test(h)) return false;
-	if (/^192\.168\./.test(h)) return false;
-	if (/^169\.254\./.test(h)) return false;
-	return true;
-}
+import { isSafePushEndpoint } from '$lib/server/ssrf';
 
 let initialized = false;
